Add TxErrorModal wip
